fix: Query cgroup memory data and fuse with /proc data for more reliable information.
The standard API for memory limits and usage under /proc are not
containerized, and return host data when used in a container. This
a source of trouble for everything trying to configure itself based
on available memory, working from bogus data.

Inside containers the cgroup information rules instead.

Trying to detect when running inside a container or not looks to have
many approaches, which are either unreliable, or non-portable, or both.

The approach here is to query both sets of data, and then fuse them for
accuracy. As we ask for the available memory taking the min of cgroup and
proc data is the safe choice for this fusion.

It assumes that inside a container the cgroup data is likely less than
the proc data, due to the container getting limited. And outside the
container the cgroup data should not be less than proc.

(That said, in the local box, having 36G memory cgroup reports a limit
of 8E (yes, exa). As proc reports the 36G correctly the min is again the
proper choice for such systems where cgroup on the host is that way out of
whack).

As for containers without limits set by docker or other runtime, cgroup
data seems to be that of the host. The min-fusion again gives the best
possible result, i.e. the host memory.